+++ EssentialFeed/EssentialFeediOSTests/Controllers/Helpers/FeedViewController+TestHelpers.swift
@@ -1,0 +1,57 @@
+//
+//  Copyright © 2019 Essential Developer. All rights reserved.
+//
+
+import UIKit
+import EssentialFeediOS
+
+extension FeedViewController {
+	func simulateUserInitiatedFeedReload() {
+		refreshControl?.simulatePullToRefresh()
+	}
+	
+	@discardableResult
+	func simulateFeedImageViewVisible(at index: Int) -> FeedImageCell? {
+		return feedImageView(at: index) as? FeedImageCell
+	}
+	
+	func simulateFeedImageViewNotVisible(at row: Int) {
+		let view = simulateFeedImageViewVisible(at: row)
+		
+		let delegate = tableView.delegate
+		let index = IndexPath(row: row, section: feedImagesSection)
+		delegate?.tableView?(tableView, didEndDisplaying: view!, forRowAt: index)
+	}
+	
+	func simulateFeedImageViewNearVisible(at row: Int) {
+		let ds = tableView.prefetchDataSource
+		let index = IndexPath(row: row, section: feedImagesSection)
+		ds?.tableView(tableView, prefetchRowsAt: [index])
+	}
+	
+	func simulateFeedImageViewNotNearVisible(at row: Int) {
+		simulateFeedImageViewNearVisible(at: row)
+		
+		let ds = tableView.prefetchDataSource
+		let index = IndexPath(row: row, section: feedImagesSection)
+		ds?.tableView?(tableView, cancelPrefetchingForRowsAt: [index])
+	}
+	
+	var isShowingLoadingIndicator: Bool {
+		return refreshControl?.isRefreshing == true
+	}
+	
+	func numberOfRenderedFeedImageViews() -> Int {
+		return tableView.numberOfRows(inSection: feedImagesSection)
+	}
+	
+	func feedImageView(at row: Int) -> UITableViewCell? {
+		let ds = tableView.dataSource
+		let index = IndexPath(row: row, section: feedImagesSection)
+		return ds?.tableView(tableView, cellForRowAt: index)
+	}
+	
+	private var feedImagesSection: Int {
+		return 0
+	}
+}
